+++ hc/front/tests/test_add_signal.py
@@ -42,6 +42,20 @@ class AddSignalTestCase(BaseTestCase):
         # Make sure it calls assign_all_checks
         self.assertEqual(c.checks.count(), 1)
 
+    def test_it_handles_username(self) -> None:
+        form = {
+            "label": "My Phone",
+            "phone": "foobar.123",
+            "down": "true",
+            "up": "true",
+        }
+
+        self.client.login(username="alice@example.org", password="password")
+        self.client.post(self.url, form)
+
+        c = Channel.objects.get()
+        self.assertEqual(c.phone.value, "foobar.123")
+
     @override_settings(SIGNAL_CLI_SOCKET=None)
     def test_it_handles_disabled_integration(self) -> None:
         self.client.login(username="alice@example.org", password="password")
@@ -72,3 +86,45 @@ class AddSignalTestCase(BaseTestCase):
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(self.url, form)
         self.assertContains(r, "Please select at least one.")
+
+    def test_it_rejects_bad_phone(self) -> None:
+        for v in ["not a phone number", False, 15, "+123456789A"]:
+            self.client.login(username="alice@example.org", password="password")
+            r = self.client.post(self.url, {"phone": v})
+            self.assertContains(r, "Invalid phone number format.")
+
+    def test_it_rejects_bad_username(self) -> None:
+        for v in ["a.123", "foobar.0"]:
+            self.client.login(username="alice@example.org", password="password")
+            r = self.client.post(self.url, {"phone": v})
+            self.assertContains(r, "Invalid username format.")
+
+    def test_it_strips_invisible_formatting_characters(self) -> None:
+        form = {"phone": "\u202c+1234567890\u202c", "down": True}
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, form)
+        self.assertRedirects(r, self.channels_url)
+
+        c = Channel.objects.get()
+        self.assertEqual(c.phone.value, "+1234567890")
+
+    def test_it_strips_hyphens(self) -> None:
+        form = {"phone": "+123-4567890", "down": True}
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, form)
+        self.assertRedirects(r, self.channels_url)
+
+        c = Channel.objects.get()
+        self.assertEqual(c.phone.value, "+1234567890")
+
+    def test_it_strips_spaces(self) -> None:
+        form = {"phone": "   +123 45 678 90   ", "down": True}
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, form)
+        self.assertRedirects(r, self.channels_url)
+
+        c = Channel.objects.get()
+        self.assertEqual(c.phone.value, "+1234567890")